requirements.txt
streamlit
litellm
chromadb
python-dotenv
jupyter
jupytext
numpy
